db tabel updated

+++ src/config/db.ts
@@ -28,7 +28,7 @@ const initDB = async () => {
       vehicle_name VARCHAR(150) NOT NULL,
       type VARCHAR(20) NOT NULL CHECK (type IN ('car','bike','van','SUV')),
       registration_number VARCHAR(100) UNIQUE NOT NULL,
-      daily_rent_price NUMERIC(10,2) NOT NULL CHECK (daily_rent_price > 0),
+      daily_rent_price INT NOT NULL CHECK (daily_rent_price > 0),
       availability_status VARCHAR(20) NOT NULL DEFAULT 'available' CHECK (availability_status IN ('available','booked')),
       created_at TIMESTAMP DEFAULT NOW(),
       updated_at TIMESTAMP DEFAULT NOW()
@@ -43,7 +43,7 @@ const initDB = async () => {
       vehicle_id INT REFERENCES vehicles(id) ON DELETE RESTRICT,
       rent_start_date DATE NOT NULL,
       rent_end_date DATE NOT NULL,
-      total_price NUMERIC(12,2) NOT NULL CHECK (total_price >= 0),
+      total_price INT NOT NULL CHECK (total_price >= 0),
       status VARCHAR(20) NOT NULL DEFAULT 'active' CHECK (status IN ('active','cancelled','returned')),
       created_at TIMESTAMP DEFAULT NOW(),
       updated_at TIMESTAMP DEFAULT NOW()
